✨ created useScroll hook
